Update RG_exploder_main.py
Fixes the incorrect definition of a 'New Variations Source' for transcripts in -1 orientation.
This happened following the change of all source DNA sequences to the +1 orientation*

a) The calculation for populating the mrnapos_lookup list was incorrect for transcripts in the -1 orientation. Modifications to begin, end, offsets to cater for this.
b) Reversing the starts and ends list is now obsolete.

Also: Some titling and comments changes.
Existing issue: the  display of Updstream/Downstream locations for -1 orientation is still incorrect

*Not spotted for some time: It was detected during web-browser version development by a failure in the App.vue script which implemented the same algorithm as here.