Update RG_exploder_main.py
Fixes the incorrect definition of a 'New Variations Source' for transcripts in -1 orientation.
This happened following the change of all source DNA sequences to the +1 orientation*

a) The calculation for populating the mrnapos_lookup list was incorrect for transcripts in the -1 orientation. Modifications to begin, end, offsets to cater for this. eg: for n in range(lookup_begin,lookup_end+1):
b) Reversing the starts and ends list is now obsolete.
Also:
c) Incomplete logic for adding the local haplotype file definitions - they were being ignored completely in favour of user-defined only.
d) get_mutfreq_from_label(inlabel) and mutfreqs_extend moved to RG_process
Some titling and comments changes.

Existing issue: the display of Upstream/Downstream locations for -1 orientation is still incorrect. Some potential code to fic this is present here.

*Clearly not properly tested at release. It was detected during web-browser version development by a failure in the App.vue script which implemented the same algorithm as here.