Data modification to identifiers for MANE_Select and change to GUI to generate modified URL for these exceptions
The transcript ID ENST00000672715, for AK2 is not defined in the GRCh37 version of Ensembl.
That's a problem because it's the MANE_Select transcript for AK2 in GRCh38, making it different from the other data in this data set, where the MANE_Select transcript exists in both GRCh37 and GRCh38
To 'fix' this two things have been done:
A) This transcript, its join positions for mRNA and CDS sequence have been taken from GRCh38 and put into the feature table of the reference sequence for AK2, but with the join positions modified to correspond to the different required offset.
These were verified by sequence alignments!! The id has had 'm' appended ie: ENST00000672715m
B) The GUI code has been modified to switch the calculated URL to point at GRCh38 for any transcript id ending with an 'm'
Voila: the "Template" link points at Ensembl GRCh38 for ENST00000672715 and at  Ensembl GRCh37 for everything else.