Modification to variant sequence assignment
With the switch to reads in +1 , the determination of the user-defined variants was incomplete, creating errors for transcripts on - strand.
There was also, unintended, reverse-complementing for non-transcript regions.
The variant definition must arrive defined as a +1 strand, whatever orientation file it goes in.
This correction makes transcript positions in 'build' show and save correctly for +1 and -1 transcripts, whatever strand.
It also makes non-transcript regions save correctly for +1 strand.
It leaves the anomaly that -1 strand sequences are not correctly handled.
There is an obvious remedy: Flip the polarity and work out how to handle thereafter. Drawback: another code-bloat for what purpose?