Update RG_exploder_main.py
A rebuild of the code identifying intron-exon boundaries. This required for:
a) Completing the switch from stored loci being sense-direction, thereby having a mixture of + and - strand source data
b) Specifically, processing differently depending on the combination of sense-strand and strand, so it can cope with source data presented in all combinations.
c) Adding an identification of pre-locus, upstream, downstream and post-locus to the structure listing.
d) Most importantly this resolved the errors in creating new variants because the feature lists were set up only for sense-strand use. All minus-strand sense transcripts were giving incorrect variant definitions. This now seems to be resolved.

NB: The big time delay was simply taking a lot of time off over the summer months !