Vue source code
These are the maintained features in a much larger set of data used to compile the python source into pyodide, wrap in Docker etc. It's not enough on its own to build the AWS "webdist" file-set (whole different set of maintenance issues there). I am guessing that the rest could be re-built but these would be a distinct reqiremenet.