Update across python source for terminology and labelling
Terminology changes around what is called: a reference; a template; a haplotype.
"absolute" for mapping position becomes "genomic"
Naming of output files for what is the reference haplotype (there were two output formats with the same name which were overwriting each other).
Journaling of the substitutions adding the local and genomic coordinates to enable cross-referencing them in the source .gb files.