Update RG_exploder_process.py
Modified the slice-function for when is_fastq_random == False (which is the new default).
When is_fastq_random == True, and read depth is high, execution time is too long, particularly in browser version because of the high number of calls to randint()

This slice tries to make amends to allow more variability in the FASTQ string without impacting execution time too much.